Number-placement Game
Write a program that places the numbers from 0 to 24 on a 5x5 table according to the rules given
below. Implement your java program using methods. Display all possible correct placements of
your numbers in your program.
 All of the numbers from 0 to 24 have been placed successfully on the board.
 Numbers must be placed in sequential order starting from 0.
 After placing a number, if there is an empty cell, the next number can be placed in the same
column or row after 2 cells or in a diagonal after 1 cell. For example;
o After the number 0 is placed randomly in the table, there are 3 possible cells for the
number 1. The number 1 can be placed two units to the right or two units down or
one unit diagonally.